Added If Not File Exists Checks.

+++ .github/custom-scripts/envato.php
@@ -24,7 +24,9 @@ TEXT;
 					'{redirect_to}',
 					'{redirect_from}',
 				), array( $item['name'], $url, $redirect_from ), $content );
-				file_put_contents( SAVE_PATH . $id . '.md', $content );
+				if ( ! file_exists( SAVE_PATH . $id . '.md' ) ) {
+					file_put_contents( SAVE_PATH . $id . '.md', $content );
+				}
 			}
 		}
 	}
